added files needed to update the database. (these will not be needed for our code at all!!!)
updating the database is so timeconsuming and as soon as i made a mistake i had to start again. hence i had some new code "cheating" which just uses the JSON of the database at any given time and puts it back into the database. this, along with kates delete all functionality which she implemented very fast, allows me to restore the database from any given point like a checkpoint system. for our MVP this is great because it means we can mess with the DB as much as we want and get anything back quickly. for my update code which didnt work many times, this saved me a lot of time when each update meant fully restoring the database. hope that makes sense - mainly a note for me :)